build-images/dagger/pyproject.toml
[project]
name = "main"
version = "0.0.0"
dependencies = [
    "jsonschema>=4.2.0",
    "PyYAML>=6.0.1"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
